requirements.txt
baidu_aip==4.16.13

Django==3.2.22

djangorestframework==3.14.0

djangorestframework_simplejwt==5.3.1

Faker==25.0.1

pypinyin==0.51.0

tencentcloud_sdk_python==3.0.1115

urllib3==1.26.15

django-simpleui

pillow

itsdangerous-2.2.0


chardet